(chore) move ActiveExercise and its subcomponents to sub-subfolder
